move race urls outside of seed url
