+++ apps/web/app/race/[key]/spoiler/download/route.ts
